Updated pizzaOrderProcessor to handle HTML form.
The servlet now prints the inout from the HTML form to the terminal.
The method used to submit this form is POST. The servlet takes the form
data, saves it to a variable and is now ready to begin producing html
with the results to the display after the HTML form is submitted.

Please note: toppings are currently not printing to the terminal. Using
multiple selection isn’t working with this variable, but the code is
not returning any errors